+++ src/librbd/ImageWatcher.h
@@ -61,7 +61,7 @@ public:
   void notify_rebuild_object_map(uint64_t request_id,
                                  ProgressContext &prog_ctx, Context *on_finish);
   void notify_rename(const std::string &image_name, Context *on_finish);
-  void notify_qos_update(int rsv, int wgt, int lmt, int bdw, Context *on_finish);
+  void notify_qos_update(int64_t rsv, int64_t wgt, int64_t lmt, int64_t bdw, Context *on_finish);
   void notify_qos_remove(int flag, Context *on_finish);
   void notify_update_features(uint64_t features, bool enabled,
                               Context *on_finish);
