+++ src/librbd/internal.h
@@ -173,9 +173,12 @@ namespace librbd {
   int rbd_cache_disable(ImageCtx *ictx);
   int rbd_cache_remove(ImageCtx *ictx);
   int qos_spec_set(ImageCtx *ictx,
-                       int rsv = -1, int wgt = -1, int lmt = -1, int bdw = -1);
+                       int64_t rsv = -1, int64_t wgt = -1, int64_t lmt = -1, int64_t bdw = -1);
   int qos_spec_get(ImageCtx *ictx,
-                       int *rsv = nullptr, int *wgt = nullptr, int *lmt = nullptr, int *bdw = nullptr,
+                       int64_t *rsv = nullptr,
+                       int64_t *wgt = nullptr,
+                       int64_t *lmt = nullptr,
+                       int64_t *bdw = nullptr,
                        int *mflag = nullptr);
   int qos_spec_del(ImageCtx *ictx, int flag = 0x0F);
 }
